[sound] alsa: hda - Update mixer name for the lower codec address
Message-id: <146374421683-129-patchset-jkysela@redhat.com>
Patchwork-id: 146595
O-Subject: [RHEL 7.3 PATCH 128/485] ALSA: hda - Update mixer name for the lower codec address
Bugzilla: 1288993
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

In most cases, we prefer the onboard codec as the primary device, thus
it's better to set it as the mixer name.  Currently, however, the
mixer name is updated per the device instantiation order, and user
gets often HDMI/DP or other seen as a mixer chip name.  Also, if a
codec name is renamed by the driver, the old chip name might be left
still as the mixer name.

This patch addresses these issues by remembering the chip address that
was referred as the mixer name.  When a codec with the same or lower
address gives its name, renew the mixer name accordingly, as it's
either the update of the codec name or we get likely the more
appropriate chip as the reference.

Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 7fbe824a0f0e6a73de5b589a6690294f7cb9dae3)
rhbz#1288993
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>