[sound] alsa: hda - Fix spurious kernel WARNING on Baytrail HDMI
Message-id: <146374421683-392-patchset-jkysela@redhat.com>
Patchwork-id: 146808
O-Subject: [RHEL 7.3 PATCH 391/485] ALSA: hda - Fix spurious kernel WARNING on Baytrail HDMI
Bugzilla: 1288993
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

snd_hdac_sync_audio_rate() call is mandatory only for HSW and later
models, but we call the function unconditionally blindly assuming that
the function doesn't do anything harmful.  But since recently, the
function checks the validity of the passed pin NID, and eventually
spews the warning if an unexpected pin is passed.  This is seen on old
chips like Baytrail.

The fix is to limit the call of this function again only for the chips
with the proper binding.  This can be identified by the same flag as
the eld notifier.

Reported-by: Ville Syrjl <ville.syrjala@linux.intel.com>
Tested-by: Ville Syrjl <ville.syrjala@linux.intel.com>
Cc: <stable@vger.kernel.org> # v4.5
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 93a9ff151754fbdf951b1b993bcf96453f6e36b3)
rhbz#1288993
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>