[netdrv] i40e/i40evf: Rewrite logic for 8 descriptor per packet check
Message-id: <1463753539-5422-136-git-send-email-sassmann@redhat.com>
Patchwork-id: 147117
O-Subject: [RHEL7.3 PATCH 135/173] i40e/i40evf: Rewrite logic for 8 descriptor per packet check
Bugzilla: 1274177
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

From: Alexander Duyck <aduyck@mirantis.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1274177
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11058163

This patch is meant to rewrite the logic for how we determine if we can
transmit the frame or if it needs to be linearized.

The previous code for this function was using a mix of division and modulus
division as a part of computing if we need to take the slow path.  Instead
I have replaced this by simply working with a sliding window which will
tell us if the frame would be capable of causing a single packet to span
several descriptors.

The logic for the scan is fairly simple.  If any given group of 6 fragments
is less than gso_size - 1 then it is possible for us to have one byte
coming out of the first fragment, 6 fragments, and one or more bytes coming
out of the last fragment.  This gives us a total of 8 fragments
which exceeds what we can allow so we send such frames to be linearized.

Arguably the use of modulus might be more exact as the approach I propose
may generate some false positives.  However the likelihood of us taking much
of a hit for those false positives is fairly low, and I would rather not
add more overhead in the case where we are receiving a frame composed of 4K
pages.

Signed-off-by: Alexander Duyck <aduyck@mirantis.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Stefan Assmann <sassmann@redhat.com>
(cherry picked from commit 2d37490b82afe1d1b745811e6ce0a4d16bc5e996)
Signed-off-by: Rafael Aquini <aquini@redhat.com>