[x86] cpu/amd: Give access to the number of nodes in a physical package
Message-id: <1463930531-23986-2-git-send-email-jolsa@redhat.com>
Patchwork-id: 147258
O-Subject: [PATCH RHEL7.3 BZ1337866 1/4] x86/cpu/amd: Give access to the number of nodes in a physical package
Bugzilla: 1337866
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1337866
https://bugzilla.redhat.com/show_bug.cgi?id=1337866

upstream
========
commit cc2749e4095cbbcb35518fb2db5e926b85c3f25f
Author: Aravind Gopalakrishnan <Aravind.Gopalakrishnan@amd.com>
Date:   Mon Jun 15 10:28:15 2015 +0200

RHEL7
=====
Using this patch as a base for following change to keep the
history straight, so we don't need the amd_get_nodes_per_socket
function.

description
===========
Stash the number of nodes in a physical processor package
locally and add an accessor to be called by interested parties.
The first user is the MCE injection module which uses it to find
the node base core in a package for injecting a certain type of
errors.

[ Rewrote the commit message, merged it with the accessor patch and unified naming. ]
Link: http://lkml.kernel.org/r/1433868317-18417-2-git-send-email-Aravind.Gopalakrishnan@amd.com

Signed-off-by: Rafael Aquini <aquini@redhat.com>