[x86] topology: Fix Intel HT disable
Message-id: <1463930531-23986-4-git-send-email-jolsa@redhat.com>
Patchwork-id: 147259
O-Subject: [PATCH RHEL7.3 BZ1337866 3/4] x86/topology: Fix Intel HT disable
Bugzilla: 1337866
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1337866
https://bugzilla.redhat.com/show_bug.cgi?id=1337866

upstream
========
commit 63d1e995be455ae9196270eb4b789de21afd42ed
Author: Peter Zijlstra <peterz@infradead.org>
Date:   Fri Mar 18 16:03:48 2016 +0100

description
===========
As per the comment in the code; due to BIOS it is sometimes impossible to know
if there actually are smp siblings until the machine is fully enumerated. So
we rather overestimate the number of possible packages.

Fixes: 1f12e32f4cd5 ("x86/topology: Create logical package id")
Link: http://lkml.kernel.org/r/20160318150538.611014173@infradead.org
Signed-off-by: Rafael Aquini <aquini@redhat.com>