[x86] perf: uncore: Remove WARN_ON_ONCE in uncore_pci_probe
Message-id: <1464021874-22119-1-git-send-email-jolsa@redhat.com>
Patchwork-id: 147313
O-Subject: [PATCH RHEL7.3 BZ1337804] perf/x86/intel/uncore: Remove WARN_ON_ONCE in uncore_pci_probe
Bugzilla: 1337804
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1337804
https://bugzilla.redhat.com/show_bug.cgi?id=1337804

upstream
========
commit ef3f00a4d38e01ec0e7ad1b1c8edc2f5667aaa32
Author: Jiri Olsa <jolsa@redhat.com>
Date:   Wed May 18 08:16:10 2016 +0200

RHEL7
=====
This bug is actually fixed by previous post of BZ1337866:
  http://post-office.corp.redhat.com/archives/rhkernel-list/2016-May/msg05183.html

but the console output also shows warning
fixed by this patch.

description
===========
When booting with nr_cpus=1, uncore_pci_probe tries to init the PCI/uncore
also for the other packages and fails with warning when they are not found.

The warning is bogus because it's correct to fail here for packages which are
not initialized. Remove it and return silently.

Fixes: cf6d445f6897 "perf/x86/uncore: Track packages, not per CPU data"
Signed-off-by: Rafael Aquini <aquini@redhat.com>