[misc] cxl: Allow initialization on timebase sync failures
Message-id: <20160523172903.48497.53218.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 147329
O-Subject: [PATCH RHEL7.3 BZ1338865 1/2] cxl: Allow initialization on timebase sync failures
Bugzilla: 1338865
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1338865

Description:
------------
Failure to synchronize the PSL timebase currently prevents the
initialization of the cxl card, thus rendering the card useless. This
is too extreme for a feature which is rarely used, if at all. No
hardware AFUs or software is currently using PSL timebase.

This patch still tries to synchronize the PSL timebase when the card
is initialized, but ignores the error if it can't. Instead, it reports
a status via /sys.

Signed-off-by: Frederic Barrat <fbarrat@linux.vnet.ibm.com>
Acked-by: Ian Munsie <imunsie@au1.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=e009a7e858fed215cb4eed5174a31cadd42d8797

----------------
Steve Best

Signed-off-by: Rafael Aquini <aquini@redhat.com>