[netdrv] ibmvnic: Enable use of multiple tx/rx scrqs
Message-id: <20160526142234.35217.59919.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 147744
O-Subject: [PATCH RHEL7.3 BZ1332848 4/4] ibmvnic: Enable use of multiple tx/rx scrqs
Bugzilla: 1332848
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1332848

Description:
------------
Enables the use of multiple transmit and receive scrqs allowing the ibmvnic
driver to take advantage of multiqueue functionality. To achieve this, the
driver must implement the process of negotiating the maximum number of
queues allowed by the server. Initially, the driver will attempt to login
with the maximum number of tx and rx queues supported by the server. If
the server fails to allocate the requested number of scrqs, it will return
partial success in the login response. In this case, we must reinitiate
the login process from the request capabilities stage and attempt to login
requesting fewer scrqs.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Upstream Status:
----------------
https://git.kernel.org/cgit/linux/kernel/git/next/linux-next.git/commit/?id=498cd8e49509c761b39dab26be7f739d95940e16

----------------
Steve Best

Signed-off-by: Rafael Aquini <aquini@redhat.com>