[fs] dlm: Replace nodeid_to_addr with kernel_getpeername
Message-id: <1464277895-23193-2-git-send-email-rpeterso@redhat.com>
Patchwork-id: 147850
O-Subject: [RHEL7.3 fs PATCH 1/2] DLM: Replace nodeid_to_addr with kernel_getpeername
Bugzilla: 1267339
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>

This patch replaces the call to nodeid_to_addr with a call to
kernel_getpeername. This avoids taking a spinlock because it may
potentially be called from a softirq context.

This is a RHEL7 port of this upstream patch:
1a31833d085a339cf42573de0717209e8b2172e8

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

rhbz#1267339

Signed-off-by: Rafael Aquini <aquini@redhat.com>