[s390] compat: correct sign-extension of the brk() compat system call
Message-id: <1464347275-11350-3-git-send-email-brueckner@redhat.com>
Patchwork-id: 148049
O-Subject: [RHEL7.3 PATCH] [s390] s390/compat: correct sign-extension of the brk() compat system call
Bugzilla: 1197172
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
The first parameter of the compat brk() system call was not
correctly sign-extended.  The reason is that the wrapper used
"llgtr" instead of "llgfr" to convert the brk() argument.

Bugzilla
--------
BZ 1197172
https://bugzilla.redhat.com/show_bug.cgi?id=1197172

Upstream status of the patch
----------------------------
The problem is RHEL-7 specific and was indirectly corrected with
the upstream commit:
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=473a06572fcd6b2e321d4b82c19ecafe383e8be9
that is available with kernel version v3.15

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=11083806

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Rafael Aquini <aquini@redhat.com>