[kernel] sched,perf: Fix periodic timers
Message-id: <1464763905-7732-8-git-send-email-jolsa@redhat.com>
Patchwork-id: 148195
O-Subject: [PATCH RHEL7.3 BZ1341230 07/44] sched,perf: Fix periodic timers
Bugzilla: 1341230
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>

Bugzilla: 1341230
https://bugzilla.redhat.com/show_bug.cgi?id=1341230

upstream
========
commit 4cfafd3082afc707653aeb82e9f8e7b596fbbfd6
Author: Peter Zijlstra <peterz@infradead.org>
Date:   Thu May 14 12:23:11 2015 +0200

RHEL7
=====
Only perf hunks.

KABI
====
Changing struct perf_cpu_context, which will consequently change
struct perf_event's KABI signature.

The 'struct perf_event*' pointer is used within 'struct thread_struct',
making perf_event struct KABI-broken any time we change it. At the moment
we dont have any perf related function on KABI list, so it's safe
to make this change.

description
===========
In the below two commits (see Fixes) we have periodic timers that can
stop themselves when they're no longer required, but need to be
(re)-started when their idle condition changes.

Further complications is that we want the timer handler to always do
the forward such that it will always correctly deal with the overruns,
and we do not want to race such that the handler has already decided
to stop, but the (external) restart sees the timer still active and we
end up with a 'lost' timer.

The problem with the current code is that the re-start can come before
the callback does the forward, at which point the forward from the
callback will WARN about forwarding an enqueued timer.

Now, conceptually its easy to detect if you're before or after the fwd
by comparing the expiration time against the current time. Of course,
that's expensive (and racy) because we don't have the current time.

Alternatively one could cache this state inside the timer, but then
everybody pays the overhead of maintaining this extra state, and that
is undesired.

The only other option that I could see is the external timer_active
variable, which I tried to kill before. I would love a nicer interface
for this seemingly simple 'problem' but alas.

Fixes: 272325c4821f ("perf: Fix mux_interval hrtimer wreckage")
Fixes: 77a4d1a1b9a1 ("sched: Cleanup bandwidth timers")
Signed-off-by: Rafael Aquini <aquini@redhat.com>