[md] clear mddev->private when it has been freed
Message-id: <1464897677-19135-11-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 148808
O-Subject: [RHEL7.3 PATCH 010/121] md: clear mddev->private when it has been freed.
Bugzilla: 1250578 1265947 1273343 1299140 1307091 1312828 1320563 1340839
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: NeilBrown <neilb@suse.de>

If ->private is set when ->run is called, it is assumed to be
a 'config'  prepared as part of 'reshape'.

So it is important when we free that config, that we also clear ->private.
This is not often a problem as the mddev will normally be discarded
shortly after the config us freed.
However if an 'assemble' races with a final close, the assemble can use
the old mddev which has a stale ->private.  This leads to any of
various sorts of crashes.

So clear ->private after calling ->free().

Reported-by: Nate Clark <nate@neworld.us>
Cc: stable@vger.kernel.org (v4.0+)
Fixes: afa0f557cb15 ("md: rename ->stop to ->free")
Signed-off-by: NeilBrown <neilb@suse.com>
(cherry picked from commit bd6919228d7e1867ae9e24ab27e3e4a366c87d21)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>