[md] close some races between setting and checking sync_action
Message-id: <1464897677-19135-19-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 148757
O-Subject: [RHEL7.3 PATCH 018/121] md: close some races between setting and checking sync_action.
Bugzilla: 1250578 1265947 1273343 1299140 1307091 1312828 1320563 1340839
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: NeilBrown <neilb@suse.com>

When checking sync_action in a script, we want to be sure it is
as accurate as possible.
As resync/reshape etc doesn't always start immediately (a separate
thread is scheduled to do it), it is best if 'action_show'
checks if MD_RECOVER_NEEDED is set (which it does) and in that
case reports what is likely to start soon (which it only sometimes
does).

So:
 - report 'reshape' if reshape_position suggests one might start.
 - set MD_RECOVERY_RECOVER in raid1_reshape(), because that is very
   likely to happen next.

Signed-off-by: NeilBrown <neilb@suse.com>
(cherry picked from commit 985ca973b68cac0adfa83497db231da7f99c6ed9)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>