[md] raid5: consider updating reshape_position at start of reshape
Message-id: <1464897677-19135-20-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 148817
O-Subject: [RHEL7.3 PATCH 019/121] md/raid5: consider updating reshape_position at start of reshape.
Bugzilla: 1250578 1265947 1273343 1299140 1307091 1312828 1320563 1340839
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: NeilBrown <neilb@suse.com>

md/raid5 only updates ->reshape_position (which is stored in
metadata and is authoritative) occasionally, but particularly
when getting closed to ->resync_max as it must be correct
when ->resync_max is reached.

When mdadm tries to stop an array which is reshaping it will:
 - freeze the reshape,
 - set resync_max to where the reshape has reached.
 - unfreeze the reshape.
When this happens, the reshape is aborted and then restarted.

The restart doesn't check that resync_max is close, and so doesn't
update ->reshape_position like it should.
This results in the reshape stopping, but ->reshape_position being
incorrect.

So on that first call to reshape_request, make sure ->reshape_position
is updated if needed.

Signed-off-by: NeilBrown <neilb@suse.com>
(cherry picked from commit 92140480ed59a006b245efd33a195fde62d1845f)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>