[md] raid5: add a new state for stripe log handling
Message-id: <1464897677-19135-45-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 148850
O-Subject: [RHEL7.3 PATCH 044/121] raid5: add a new state for stripe log handling
Bugzilla: 1250578 1265947 1273343 1299140 1307091 1312828 1320563 1340839
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Shaohua Li <shli@fb.com>

When a stripe finishes construction, we write the stripe to raid in
ops_run_io normally. With log, we do a bunch of other operations before
the stripe is written to raid. Mainly write the stripe to log disk,
flush disk cache and so on. The operations are still driven by raid5d
and run in the stripe state machine. We introduce a new state for such
stripe (trapped into log). The stripe is in this state from the time it
first enters ops_run_io (finish construction) to the time it is written
to raid. Since we know the state is only for log, we bypass other
check/operation in handle_stripe.

Signed-off-by: Shaohua Li <shli@fb.com>
Signed-off-by: NeilBrown <neilb@suse.com>
(cherry picked from commit b70abcb24711d1327a8a505ab3e931c24cbab0a7)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>