[redhat] Build kernel helper binaries with debug info
Message-id: <1464984534-15871-1-git-send-email-plougher@redhat.com>
Patchwork-id: 149050
O-Subject: [RHEL7.3 PATCH] Build kernel helper binaries with debug info
Bugzilla: 1171988
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1171988
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11122755
Upstream status: RHEL specific

For many years it has been impossible to install debuginfo packages
for different kernel versions simultaneously due to build-id conflicts.
For almost as long, this has been blamed on the fact that the source
path embedded in the binaries does not contain the KVRA.  The source
path is used in the computation of the build-id, and by definition, if
it includes the KVRA then the build-id for different kernel versions
will be different, even if nothing else has changed (source code etc.).

However, looking at the embedded source pathname in the binaries
before and after debugedit is run when building the kernel packages,
shows they do have the KVRA, and so something else is causing the
build-id conflicts.  Debugedit is used to change the real source
pathname to have the prefix "/usr/src/debug", it then recalculates the
build-id.

Analysing a substantial amount of kernel debuginfo packages, and the
arising build-id conflicts, show they all prove to be from the kernel
scripts directory, these are largely kernel build time helper binaries.
They are not packaged within the main kernel rpm, but are included in
the kernel-devel package.

So there's something wrong with the script binaries.

Grepping a sample script binary for the embedded source pathname shows
nothing, in fact there's no debug information at all.  It doesn't take
a great leap to realise the script binaries are not being compiled
with -g, and that's why they have no source pathname with the KVRA, and
that's why their build-ids conflict.

Looking in the top-level Makefile we see

HOSTCFLAGS   = -Wall -Wmissing-prototypes -Wstrict-prototypes -O2 -fomit-frame-pointer

Testing: Adding the -g flag and rebuilding the kernels with conflicting
build-ids, show the build-id conflicts go away.

Plus the *.debug files for the script binaries in the debuginfo package show
the presence of the embedded source pathname.

[plougher@lenovo-rd230-04 scripts]$ objdump -W conmakehash.debug
[SNIP]
Contents of the .debug_info section:

  Compilation Unit @ offset 0x0:
   Length:        0xc32 (32-bit)
   Version:       4
   Abbrev Offset: 0
   Pointer Size:  8
 <0><b>: Abbrev Number: 1 (DW_TAG_compile_unit)
    < c>   DW_AT_producer    : (indirect string, offset: 0x70): GNU C 4.8.5 20150623 (Red Hat 4.8.5-4) -mtune=generic -march=x86-64 -g -O2 -std=gnu90 -fomit-frame-pointer
    <10>   DW_AT_language    : 1        (ANSI C)
    <11>   DW_AT_name        : (indirect string, offset: 0x48): scripts/conmakehash.c
    <15>   DW_AT_comp_dir    : (indirect string, offset: 0x2f5): /usr/src/debug/kernel-3.10.0-420.el7.bz1171988/linux-3.10.0-420.el7.bz1171988.x86_64
[SNIP]

Signed-off-by: Phillip Lougher <plougher@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>