[scsi] hpsa: add in sas transport class
Message-id: <1464987242-38163-37-git-send-email-jszczype@redhat.com>
Patchwork-id: 149091
O-Subject: [RHEL7.3 e-stor PATCH 36/48] hpsa: add in sas transport class
Bugzilla: 1274467
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Kevin Barnett <kevin.barnett@pmcs.com>

The cherry-pick of this commit had a conflict with a
RHEL-only commit done for RHEL7.2.   It was resolved
by applying this commit in way that kept the changes
done for RHEL7.2, in hpsa_slave_alloc(), while adding
in the upstream changes to this function.

This upstream commit is:
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=d04e62b9d63a7498735761dc40eaed88b7fd9d80

RHEL7.2 commit is:
[scsi] hpsa: add support for tagged queueing
http://git.app.eng.bos.redhat.com/git/rhel7.git/commit/?id=dd3de195a00c52820176df86ce9319ddfabcdced

Remaining text is from original upstream commit:

Reviewed-by: Scott Teel <scott.teel@pmcs.com>
Reviewed-by: Justin Lindley <justin.lindley@pmcs.com>
Reviewed-by: Kevin Barnett <kevin.barnett@pmcs.com>
Reviewed-by: Hannes Reinecke <hare@suse.de>
Reviewed-by: Tomas Henzl <thenzl@redhat.com>
Reviewed-by: Matthew R. Ochs <mrochs@linux.vnet.ibm.com>
Signed-off-by: Don Brace <don.brace@pmcs.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>